Update MainActivity.cs
Correcting attribute MetaData name for AccessibilityService. This will fix the registration of metadata. This may save someone much time in troubleshooting why, for example, e.Source is ALWAYS null (canRetrieveWindowContent="true" also needs to be set in the accessibility_service_config.xml file, which is it in this codebase) - because the metadata registration was incorrect!